Moved the central nodeKey into the Context object.
