+++ thunder-core/src/main/java/network/thunder/core/communication/nio/ServerMain.java
@@ -1,6 +1,9 @@
 package network.thunder.core.communication.nio;
 
-import network.thunder.core.communication.objects.p2p.PubkeyIPObject;
+import network.thunder.core.communication.objects.p2p.sync.PubkeyIPObject;
+import org.bitcoinj.core.ECKey;
+
+import java.math.BigInteger;
 
 /**
  * Created by matsjerratsch on 19/10/2015.
@@ -9,7 +12,10 @@ public class ServerMain {
 
     public static void main (String[] args) throws Exception {
 
-        P2PContext context = new P2PContext(8993);
+        P2PContext context = new P2PContext(8992);
+
+        ECKey key = ECKey.fromPrivate(BigInteger.ONE.multiply(BigInteger.valueOf(10000)));
+        context.nodeKey = key;
 
         PubkeyIPObject ip = new PubkeyIPObject();
         ip.IP = "127.0.0.1";
@@ -19,7 +25,7 @@ public class ServerMain {
         PubkeyIPObject ip2 = new PubkeyIPObject();
         ip2.IP = "127.0.0.1";
         ip2.port = 8993;
-        context.IPList.add(ip2);
+//        context.IPList.add(ip2);
         context.startConnections();
 
     }
